refactor: User Info Edit Component
